move constants to frontend folder for containers, add docker files and built react app. some styling tweaks
