--- constants/games.js
@@ -1,59 +1,0 @@
-export const GAMES = [
-  {
-      name: "Sven Coop",
-      type: "protocol-valve",
-      description: "Sven Coop is a cooperative multiplayer game based on the Half-Life modding engine. It is a game where players work together to complete various missions and challenges.",
-      image: "https://example.com/sven-coop.jpg",
-  },
-  {
-      name: "Half-Life Deathmatch (Steam)",
-      type: "protocol-valve",
-      description: "Half-Life Deathmatch (Steam) is a multiplayer game based on the Half-Life modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/half-life-deathmatch.jpg",
-  },
-  {
-      name: "Half-Life 2 DM",
-      type: "protocol-valve",
-      description: "Half-Life 2 DM is a multiplayer game based on the Half-Life 2 modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/half-life-2-dm.jpg",
-  },
-  {
-      name: "UT99",
-      type: "ut",
-      description: "UT99 is a multiplayer game based on the Unreal Tournament modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/ut99.jpg",
-  },
-  {
-      name: "BF2",
-      type: "bf2",
-      description: "BF2 is a multiplayer game based on the Battlefield 2 modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/bf2.jpg",
-  },
-  {
-      name: "GoldenEye Source",
-      type: "protocol-valve",
-      description: "GoldenEye Source is a multiplayer game based on the GoldenEye modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/goldeneye-source.jpg",
-  },
-  {
-      name: "Alien Swarm Reactive Drop",
-      type: "protocol-valve",
-      description: "Alien Swarm Reactive Drop is a multiplayer game based on the Alien Swarm modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/alien-swarm-reactive-drop.jpg",
-  },
-  {
-      name: "WRECKFEST",
-      description: "WRECKFEST is a multiplayer game based on the WRECKFEST modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/wreckfest.jpg",
-  },
-  {
-      name: "Quake 3",
-      description: "Quake 3 is a multiplayer game based on the Quake 3 modding engine. It is a game where players compete against each other in various missions and challenges.",
-      image: "https://example.com/quake-3.jpg",
-  },
-  {
-      name: "Steam ID Finder",
-      description: "Steam ID Finder is a tool that allows you to find the Steam ID of a player. It is a tool that is used to find the Steam ID of a player.",
-      image: "https://example.com/steam-id-finder.jpg",
-  },
-];
