+++ frontend/src/constants/gameServerInfo.js
