+++ frontend/src/hooks/useServerData.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import io from 'socket.io-client';
 
-const SOCKET_URL = 'http://localhost:3010';
+const SOCKET_URL = import.meta.env.VITE_SOCKET_URL || 'http://localhost:3010';
 
 export function useServerData() {
   const [data, setData] = useState(null);
