+++ frontend/src/pages/ServerConfig.jsx
@@ -1,5 +1,5 @@
 import Card from '../components/Card';
-import { gameServerInfo } from '../../../constants/gameServerInfo'; 
+import { gameServerInfo } from '../constants/gameServerInfo'; 
 
 const ServerConfig = () => {
   return (
